+++ internal/libtaskido/printer_test.go
@@ -7,28 +7,28 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestApplyColorToDate(t *testing.T) {
+func TestFormatDate(t *testing.T) {
 	tests := []struct {
 		date        string
 		expectation string
 	}{
 		// Two days after date
-		{time.Now().Add(-48 * time.Hour).Format("2006-01-02"), "\033[0;31m" + time.Now().Add(-48*time.Hour).Format("2006-01-02") + "\033[0;0m"},
+		{time.Now().Add(-48 * time.Hour).Format("2006-01-02"), "\033[0;31m⚑\033[0;0m "+time.Now().Add(-48*time.Hour).Format("2006-01-02")},
 		// Yesterday date
-		{time.Now().Add(-24 * time.Hour).Format("2006-01-02"), "\033[0;31m" + time.Now().Add(-24*time.Hour).Format("2006-01-02") + "\033[0;0m"},
+		{time.Now().Add(-24 * time.Hour).Format("2006-01-02"), "\033[0;31m⚑\033[0;0m "+time.Now().Add(-24*time.Hour).Format("2006-01-02")},
 		// Today date
-		{time.Now().Truncate(24 * time.Hour).Format("2006-01-02"), "\033[0;31m" + time.Now().Truncate(24*time.Hour).Format("2006-01-02") + "\033[0;0m"},
+		{time.Now().Truncate(24 * time.Hour).Format("2006-01-02"), "\033[0;31m⚑\033[0;0m "+ time.Now().Truncate(24*time.Hour).Format("2006-01-02")},
 		// Tomorrow date
-		{time.Now().Add(24 * time.Hour).Format("2006-01-02"), "\033[0;33m" + time.Now().Add(24*time.Hour).Format("2006-01-02") + "\033[0;0m"},
+		{time.Now().Add(24 * time.Hour).Format("2006-01-02"), "\033[0;33m⚑\033[0;0m "+time.Now().Add(24*time.Hour).Format("2006-01-02")},
 		// Two day before date
-		{time.Now().Add(48 * time.Hour).Format("2006-01-02"), "\033[0;32m" + time.Now().Add(48*time.Hour).Format("2006-01-02") + "\033[0;0m"},
+		{time.Now().Add(48 * time.Hour).Format("2006-01-02"), "  "+time.Now().Add(48*time.Hour).Format("2006-01-02")},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.date, func(t *testing.T) {
-			result := applyColorToDate(tt.date)
+			result := formatDate(tt.date)
 			if result != tt.expectation {
-				t.Errorf("applyColorToDate(%s) = %s; want %s", tt.date, result, tt.expectation)
+				t.Errorf("formatDate(%s) = %s; want %s", tt.date, result, tt.expectation)
 			}
 		})
 	}
@@ -89,7 +89,7 @@ func TestFormatTask(t *testing.T) {
 				Archived:  false,
 				Priority:  0,
 			},
-			expected: []string{"1", "\033[0;31m2024-07-24\033[0;0m", "\033[0;35mproject1\033[0;0m", "\033[0;35mproject2\033[0;0m", "\033[0;34m@context1\033[0;0m", "Do", "something"},
+			expected: []string{"1", "\033[0;31m⚑\033[0;0m", "2024-07-24", "\033[0;35mproject1\033[0;0m", "\033[0;35mproject2\033[0;0m", "\033[0;34m@context1\033[0;0m", "Do", "something"},
 		},
 		{
 			task: Task{
@@ -101,7 +101,7 @@ func TestFormatTask(t *testing.T) {
 				Archived:  false,
 				Priority:  3,
 			},
-			expected: []string{"2", "\033[0;31m\u278C\033[0;0m", "\033[0;31m2024-07-23\033[0;0m", "\033[0;35mproject3\033[0;0m", "\033[0;34m@context2\033[0;0m", "Another", "task"},
+			expected: []string{"2", "\033[0;31m\u278C\033[0;0m", "\033[0;31m⚑\033[0;0m", "2024-07-23", "\033[0;35mproject3\033[0;0m", "\033[0;34m@context2\033[0;0m", "Another", "task"},
 		},
 	}